requirements_torch.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1658854572623/work
aiohttp @ file:///D:/bld/aiohttp_1649013334059/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1636093929600/work
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
astunparse==1.6.3
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1659030588191/work
bleach==4.1.0
blinker==1.4
brotlipy @ file:///D:/bld/brotlipy_1648854327487/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1640686991047/work
certifi==2022.6.15
cffconvert==2.0.0
cffi @ file:///D:/bld/cffi_1656782930891/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
click @ file:///D:/bld/click_1651215430313/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1653061851209/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
commonmark==0.9.1
cryptography @ file:///D:/bld/cryptography_1657174166005/work
csbdeep==0.7.2
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
cytoolz @ file:///D:/bld/cytoolz_1657553547471/work
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1658528993411/work
diplib==3.3.0
docopt==0.6.2
elf @ file:///home/conda/feedstock_root/build_artifacts/python-elf_1653757506874/work
et-xmlfile==1.0.1
flatbuffers==1.12
fonttools @ file:///D:/bld/fonttools_1657249503498/work
frozenlist @ file:///D:/bld/frozenlist_1648771863087/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1659099796968/work
gast==0.4.0
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1657692258949/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio @ file:///D:/bld/grpc-split_1658277577368/work
h5py @ file:///D:/bld/h5py_1656336270487/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imagecodecs @ file:///D:/bld/imagecodecs_1659059254725/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1654071881944/work
importlib-metadata @ file:///D:/bld/importlib-metadata_1653252997230/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1633637554808/work
jsonschema==3.2.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver @ file:///D:/bld/kiwisolver_1657953200729/work
libclang==14.0.6
llvmlite==0.38.1
locket @ file:///home/conda/feedstock_root/build_artifacts/locket_1650660393415/work
Mako @ file:///home/conda/feedstock_root/build_artifacts/mako_1656619049246/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1657921053229/work
MarkupSafe @ file:///D:/bld/markupsafe_1648737757460/work
matplotlib @ file:///D:/bld/matplotlib-suite_1659031692802/work
mrcfile @ file:///home/conda/feedstock_root/build_artifacts/mrcfile_1657824357067/work
multidict @ file:///D:/bld/multidict_1648882567277/work
munkres==1.1.4
natsort==8.1.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1658215017934/work
numba @ file:///D:/bld/numba_1655473409142/work
numpy @ file:///D:/bld/numpy_1653325517362/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opencv-python==4.6.0.66
openpyxl @ file:///home/conda/feedstock_root/build_artifacts/openpyxl_1632332968451/work
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas @ file:///D:/bld/pandas_1656001279371/work
partd @ file:///home/conda/feedstock_root/build_artifacts/partd_1617910651905/work
patchify==0.2.3
Pillow @ file:///D:/bld/pillow_1657007292472/work
pkginfo==1.8.2
plantseg @ file:///usr/share/miniconda/conda-bld/plantseg_1644503641058/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1657729053205/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyclesperanto-prototype @ file:///home/conda/feedstock_root/build_artifacts/pyclesperanto-prototype_1659888471990/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1652398519695/work
pykwalify==1.8.0
pyopencl==2022.1.6
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///D:/bld/pysocks_1648857426942/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==0.20.0
pytools @ file:///home/conda/feedstock_root/build_artifacts/pytools_1656247651513/work
pytorch3dunet==1.3.7
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1647961439546/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets @ file:///D:/bld/pywavelets_1649616575120/work
pywin32-ctypes==0.2.0
PyYAML @ file:///D:/bld/pyyaml_1648757346451/work
readme-renderer==34.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.1.0
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image @ file:///D:/bld/scikit-image_1655254103169/work
scikit-learn @ file:///D:/bld/scikit-learn_1652976858669/work
scipy @ file:///C:/bld/scipy_1658811088396/work
segmentation-models-3D==1.0.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
skan @ file:///home/conda/feedstock_root/build_artifacts/skan_1637843622441/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1651519171059/work/tensorboard-2.9.0-py3-none-any.whl
tensorboard-data-server @ file:///D:/bld/tensorboard-data-server_1649932990984/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1659079987845/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==1.9.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1649051611147/work
transforms3d==0.3.1
twine==4.0.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
unicodedata2 @ file:///D:/bld/unicodedata2_1649112114189/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
vollseg==5.0.8
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1658974769507/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1648771910527/work
wrapt==1.14.1
yacs==0.1.8
yarl @ file:///D:/bld/yarl_1648966682693/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
